Initial import of map display for locations.
NOTE: At the time this commit was made, an error in bundler dealing
with es6 nullish coalescing operator in the react leaflet caused the
build to fail with an "unexpected token".

https://github.com/PaulLeCam/react-leaflet/issues/877

The only way I could continue developing was to install an older version
of react leaflet. Once a real solution is available, we can bump up
to the latest version.

This commit is mainly to layout groundwork of what the class signup
page could look like. For development sake, it's been placed on the
homepage with fake data.