Allow location extraction from params.
One of the objectives of the new signup with location page is to
allow for sharable links with the location and experience level
already selected. This adds a query extractor that checks for any
params that can then be fed into the page (for the Title) and fetch
the correct classes when inserted into the course offering endpoint.