Adds 'extras' field
We're walking a tight rope of having some data live here and some in the
PP dashboard.

As the PP dashboard holds the class inventory, we figure pulling it from
there makes sense.

I named 'extras' rather than 'just' semesters to give clarity that this
stuff is comes from a special case and because I expect it to pull in
more 'context' overtime

We're planning on having there be a regular chron job to rebuild this
site so this data stays up to date.