Extract filterTemplate and sortSemester.
Locations don't know about semesters and vis versa. This creates a data
mismatch when filtering by semesters that makes it appear as though the
same class category options exist on both semesters. Until we have
a way to relate this data, it's best to hide the semester filter to
avoid confusion.

We extract the filterTemplate and pass it via prop to determine which
filters should be displayed now. This also required that sortSemester
be shared as a helper.