Update to Gatsby 3.
We apparently used an outdated version of gatsby when we first
started this project, and it hasn't been an issue until storybook
brought some newer dependencies like webpack 5.

This change brings significant loading time and performance improvements,
an overhauled image plugin, improved graphql handling.

This commit also includes a fix for a breaking change that did away
with fmImagesToRelative preprocessing in the gastby build stage.