Ensure storybook plays nice with SCSS.
Components have had a blend of styling in multiple locations sometimes
dependent on the main 'all.scss' file and others directly embedded in
the styles prop of the component. While this was nice having components
self contained in their file, it's led to some discrepancies in style
as it was difficult determining where the style was coming from.

To address this, only styles that can be changed via the page builder
live on the component, while general component styling will be in
corresponding scss files.