First story: ButtonLink.
Extracted from the Buttons component to be an independent atom for
use in other parts of the app.

This introduces documentation via the props that automatically gets
rendered in the description of the story book for clarity in what
the props do.

This is particularly useful for components related to PageBuilder because
I had to standardize the props labels for the graphql to function
across multiple components.