Make props consistent for Reviews.
Due to the standard prop names necessary for the graphQL to work with
the page builder, the names don't match the function in the components.
They were renamed to make more sense, but it's also more difficult to
track for the programmer. This makes the names consistent and ensures
the documentation clearly states what the prop is used for.