Include phone number in locations.
I dislike that these are hard-coded and embedded at build time, but this
is a temporary fix to including phone numbers on the location pages.

The reason this isn't a full fledge solution is that we're looking into
creating a more robust way of dealing with location pages that will
allow for independent customization of location details and programs
that a market is running.