Add data structure of side by side in CMS.
The replicates the functionality of page builder allowing a user to
select from existing components as types. One limitation is that this
component only expects one left and one right component, however the
CMS allows for an array of choices. In the component, only the first
in each array is used.