Add Iframe component for interactivity.
One of the biggest selling points of TCS is the ability for kids to
create cool projects and it's very important that we can display these
on our pages. This introduces a new iframe component that allows the
embedding of projects, maps, etc.

This will be added to the PageBuilder and potentially separately in
a side by side component.