Register new iframe specific fields in graphql.
It's important that graphql has an example of what the fields look like,
otherwise it cannot infer what the field should be. In this case, a
title, content, and a size object have been introduces to the
graphQL for custom pages. This will properly extract the data from
the page markdown.

Interesting to note: it appears that registering in one graphQL query
seems to be enough to apply to other pages.