Adds custom location data to CMS and query.
We add location pages at build time to have the most up to date information
about class locations pulled directly from the Dashboard. This data
was automatically added to a location template. While useful for
data integrity, it handicapped the ability to customize information
on a per market basis.

Markets want more say so in what they're advertising or might want to
share information only relevant to that location. This new addition
of custom location page data makes room for that ability. Data is stored
in new locations prefixed with "custom-" to avoid conflicts with the
pages created at build time. This data is then queried in the location
page template, and will be inserted into the page.