Add custom location page preview template.
Due to the complexity of the real location page importing data at build
time to create the page, we cannot use the same template as the real page
and cannot access the location specific data as previews don't have
access to graphql data.

Therefore we've created a basic preview template to show what data can
be customized and we use placeholder data for the remainder.