Setup empty structure of homepage.
Outline of the content of the homepage and moves over existing
functionality to new structure.

First, the hero section contains the eye-catching video and immediate
links to signup for classes or camps. This is intended to take up the
full visible screen for a clean look. Other helpful 'jump-to' links
will be added for easier navigation of the homepage.

Second, the info section will contain the main pitch for the site and
other information that sets TCS apart.

Third, I wanted the site to provide specific examples of what TCS is
and what they should expect from their students. I think that it would
be helpful to display a featured project section that shows an
interactive iframe project and info about the student and their project.
Winners of competitions can be featured here as an added bonus!
Additionally, a featured class will show a picture and description of
a rotating set of classes with specific details about what kids
will learn (skills, projects). Lastly, featured blogs will show recent
or important blogs on our site.

Fourth, testimonials and partnered schools will be displayed at the
bottom to show legitimacy of our offerings.