IOS date picker fix for 0 height/width bug.
Only on mobile IOS devices, the default date picker is rendering
with 0 height and width, causing the field to look flat with the
perceived impression that it is unclickable. This forces a minimum
height and width on the fields so that it looks clickable.

Once the user clicks, the field works as expected.