Add support for instant preview of uploaded imgs.
Our array of team members have an image object for each member. In order
to see the preview of an image that was just uploaded, we have to get
the asset in each member and that returns a url that can be processed
as the src on an img tag.