Temp change config branch for image testing.
Improvements to our image handling will include the ability to provide
a URL as an image source rather than hosting on our site (this is only
advised for things like blogs).

This will reduce the clutter of the images stored on our site and can
be really helpful for blog posts that list products or people.