Retain URL integrity by using default slugs.
Netlify/Gatsby already use the title as the default for creating slugs,
and has no built in way to keep the slug/file md name/title in sync when
it's changed. This is intentional so that links to break! Therefore,
we'll leave the ability to edit the slug unchanged for now to avoid any
issues in the future.

In this case I've manually updated the file name to match the title, but
going forward it makes more sense to 'archive' an unused experience level
rather than edit an existing one. This is now a essentially a
should display field on the experience level cms.