Restore placeholder custom page for graphql query.
Left and Right side by side have properties declared on this custom page
that are necessary at the moment to allow the pageBuilder query to
work properly, as graphql can't infer the types without having an
example in a markdown file. In the future, this can likely be removed
or we can define the schema in a separate file.