Add vaccinated question to overview page.
Starting January 31, 2022, all students are required to have at least
one dose of a vaccine in order to participate in in-person classes.I've
added a friendly alert box to the top of the overview announcing this
policy, and added a temporary formik radio group to take input.

A customOnChange handles the response, allowing form submission if the
the user has confirmed vaccination status or if the class is online. If
the user marks no or tries to skip the question, we block submission of
the form and suggest the user try an online class.

This question doesn't need to be processed on our backend as it should
be safe to assume all students have been vaccinated if they've made it
to the submission step.