Configure honeybadger util for notify.
The gastby plugin for honeybadger will notify HB if any errors occur
on the site, however it does not give us access to notify a custom
error. In this case we want to notify HB anytime a user encounters
a 404 error. This is especially important in the website transition
to catch broken links.