Remove Honeybadger 404 notification.
This added a lot of noise to our Honeybadger account and was really only
useful during the transition to the new site in early 2022. If a page is
missing at this point, it's better to notify staff and fix it.

Most of the Honeybadger errors are from old internet scrapers a link
checkers