Adds "dummy content" for graphql to be happy.
GraphQL either needs the data to be defined in a schema like structure,
or it makes assumptions based on existing markdown data. In this case,
Maddy deleted the mediaPosition and list fields from one of the program
pages, and it caused issues when graphQL was loading and inferring the
data types.

Creating a schema is an undertaking we're not ready for, so in the
meantime we can add "dummy content" that doesn't actually get
rendered.